requirements.txt
Django~=4.1
django-phone-field==1.8.1
django-phonenumber-field
django-phonenumbers
